at branch: move AttendanceItemType to shared, add AttendanceItemType webservice

+++ nts.uk/ctx/at/schedule/nts.uk.ctx.at.schedule.app/src/main/java/nts/uk/ctx/at/schedule/app/command/budget/premium/UpdatePremiumItemCommand.java
@@ -8,13 +8,13 @@ import lombok.Value;
 public class UpdatePremiumItemCommand {
 	private String companyID;
 	
-	private BigDecimal iD;
+	private Integer iD;
 	
-	private BigDecimal attendanceID;
+	private Integer attendanceID;
 	
 	private String name;
 
-	private BigDecimal displayNumber;
+	private Integer displayNumber;
 
 	private int useAtr;
 }
