+++ nts.uk/ctx/at/schedule/nts.uk.ctx.at.schedule.app/src/main/java/nts/uk/ctx/at/schedule/app/find/budget/premium/PremiumItemDto.java
@@ -12,13 +12,13 @@ import nts.uk.ctx.at.schedule.dom.budget.premium.UseAttribute;
 public class PremiumItemDto {
 	private String companyID;
 	
-	private BigDecimal iD;
+	private Integer iD;
 	
-	private BigDecimal attendanceID;
+	private Integer attendanceID;
 	
 	private String name;
 
-	private BigDecimal displayNumber;
+	private Integer displayNumber;
 
 	private int useAtr;
 }
